Move random_state to utils
