+++ src/qibo/tests_new/test_core_gates_features.py
@@ -4,7 +4,7 @@ import numpy as np
 import qibo
 from qibo import K, gates
 from qibo.models import Circuit
-from qibo.tests_new.test_core_gates import random_state
+from qibo.tests_new.utils import random_state
 
 
 ####################### Test `construct_unitary` feature #######################
